pyproject.toml
[tool.poetry]
name = "temporal-lib-py"
version = "1.3.2"
description = "A wrapper library for candid-based temporal authentication"
authors = ["gtato"]
readme = "README.md"
packages = [{include = "temporallib"}]
license = "LGPL-3.0"

[tool.poetry.dependencies]
python = "^3.8"
macaroonbakery = "^1.3.1"
temporalio = "^1.3.0"
pycryptodome = "^3.15.0"
google-auth = "^2.19.1"
sentry-sdk = "^1.29.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.8.0"
isort = "^5.10.1"
poethepoet = "^0.16.2"
pytest-asyncio = "^0.19.0"
mypy = "^0.971"

[tool.poe.tasks]
format = [{cmd = "black ."}, {cmd = "isort ."}]
lint = [{cmd = "black --check ."}, {cmd = "isort --check-only ."}, {ref = "lint-types" }]
lint-types = "mypy --check-untyped-defs ."
test = "pytest"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"

[tool.isort]
profile = "black"
skip_gitignore = true

[tool.mypy]
ignore_missing_imports = true
